set jwt as cookie in login, use cookie for authorization

+++ api/src/main/java/com/tvt21webdev2/climatechangecharts/controller/ViewController.java
@@ -2,6 +2,7 @@ package com.tvt21webdev2.climatechangecharts.controller;
 
 import java.util.List;
 
+import com.tvt21webdev2.climatechangecharts.service.SecurityService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -14,9 +15,11 @@ import com.tvt21webdev2.climatechangecharts.service.ViewService;
 public class ViewController {
 
   private final ViewService service;
+  private final SecurityService securityService;
 
-  public ViewController(final ViewService service) {
+  public ViewController(final ViewService service, final SecurityService securityService) {
     this.service = service;
+    this.securityService = securityService;
   }
 
   @GetMapping("/view")
@@ -27,14 +30,21 @@ public class ViewController {
   }
 
   @PostMapping("/view/create")
-  public ResponseEntity<String> createView(@RequestBody View view) {
+  public ResponseEntity<String> createView(@RequestBody View view, @CookieValue(name = "token") String token) {
+    String username = securityService.validateJwt(token);
+    if (securityService.validateJwt(username) == null) {
+      return new ResponseEntity<>("Token not valid", HttpStatus.UNAUTHORIZED);
+    }
     service.saveView(view);
     return new ResponseEntity<>(view.getUrl() + " created successfully", HttpStatus.OK);
   }
 
   @PostMapping("/view/delete")
-  public ResponseEntity<String> deleteView(@RequestParam(defaultValue = "empty") String id) {
-    
+  public ResponseEntity<String> deleteView(@RequestParam(defaultValue = "empty") String id, @CookieValue(name = "token") String token) {
+    String username = securityService.validateJwt(token);
+    if (securityService.validateJwt(username) == null) {
+      return new ResponseEntity<>("Token not valid", HttpStatus.UNAUTHORIZED);
+    }
     service.deleteById(Long.parseLong(id));
     return new ResponseEntity<>("View deleted", HttpStatus.OK);
   }
